working cvx shim

+++ coneOSsparse/util.h
@@ -1,14 +1,13 @@
 #ifndef UTIL_H_GUARD
 #define UTIL_H_GUARD
 
-#include <time.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include "coneOS.h"
 
 void tic(void); 
-float toc(void); 
-float tocq(void); 
+double toc(void); 
+double tocq(void); 
 void printConeData(Cone * k);
 void printData(Data * d);
 void printAll(Data * d, Work * w);
